Move `DiagnosticMessage` into separate module
